refactor: add missing exports, better error handling, and handle boolean query params
