+++ requirements-dev.txt
@@ -4,17 +4,17 @@
 #
 #    pip-compile requirements-dev.in
 #
-asgiref==3.9.1
+annotated-types==0.7.0
+    # via pydantic
+asgiref==3.9.2
     # via django
 asv==0.6.5
     # via -r requirements-dev.in
 asv-runner==0.2.1
     # via asv
-backports-asyncio-runner==1.2.0
-    # via pytest-asyncio
 build==1.3.0
     # via asv
-cachetools==6.1.0
+cachetools==6.2.0
     # via tox
 certifi==2025.8.3
     # via requests
@@ -28,7 +28,7 @@ colorama==0.4.6
     # via tox
 distlib==0.4.0
     # via virtualenv
-django==4.2.24
+django==5.2.6
     # via
     #   -r requirements-dev.in
     #   django-template-partials
@@ -36,15 +36,13 @@ django-template-partials==25.2
     # via -r requirements-dev.in
 djc-core-html-parser==1.0.2
     # via -r requirements-dev.in
-exceptiongroup==1.3.0
-    # via pytest
 filelock==3.19.1
     # via
     #   tox
     #   virtualenv
 greenlet==3.2.4
     # via playwright
-identify==2.6.13
+identify==2.6.14
     # via pre-commit
 idna==3.10
     # via requests
@@ -52,7 +50,6 @@ importlib-metadata==8.7.0
     # via
     #   asv
     #   asv-runner
-    #   build
 iniconfig==2.1.0
     # via pytest
 json5==0.12.1
@@ -74,11 +71,11 @@ pathspec==0.12.1
     # via
     #   -r requirements-dev.in
     #   mypy
-platformdirs==4.3.8
+platformdirs==4.4.0
     # via
     #   tox
     #   virtualenv
-playwright==1.54.0
+playwright==1.55.0
     # via -r requirements-dev.in
 pluggy==1.6.0
     # via
@@ -86,6 +83,10 @@ pluggy==1.6.0
     #   tox
 pre-commit==4.3.0
     # via -r requirements-dev.in
+pydantic==2.11.9
+    # via -r requirements-dev.in
+pydantic-core==2.33.2
+    # via pydantic
 pyee==13.0.0
     # via playwright
 pygments==2.19.2
@@ -101,52 +102,45 @@ pyproject-api==1.9.1
     # via tox
 pyproject-hooks==1.2.0
     # via build
-pytest==8.4.1
+pytest==8.4.2
     # via
     #   -r requirements-dev.in
     #   pytest-asyncio
     #   pytest-django
     #   syrupy
-pytest-asyncio==1.1.0
+pytest-asyncio==1.2.0
     # via -r requirements-dev.in
 pytest-django==4.11.1
     # via -r requirements-dev.in
-pyyaml==6.0.2
+pyyaml==6.0.3
     # via
     #   asv
     #   pre-commit
-requests==2.32.4
+requests==2.32.5
     # via -r requirements-dev.in
 ruff==0.13.2
     # via -r requirements-dev.in
 sqlparse==0.5.3
     # via django
-syrupy==4.9.1
+syrupy==5.0.0
     # via -r requirements-dev.in
 tabulate==0.9.0
     # via asv
-tomli==2.2.1
-    # via
-    #   asv
-    #   build
-    #   mypy
-    #   pyproject-api
-    #   pytest
-    #   tox
-tox==4.28.4
+tox==4.30.2
     # via -r requirements-dev.in
-types-requests==2.32.4.20250809
+types-requests==2.32.4.20250913
     # via -r requirements-dev.in
-typing-extensions==4.14.1
+typing-extensions==4.15.0
     # via
     #   -r requirements-dev.in
-    #   asgiref
-    #   exceptiongroup
     #   mypy
+    #   pydantic
+    #   pydantic-core
     #   pyee
     #   pytest-asyncio
-    #   tox
-    #   virtualenv
+    #   typing-inspection
+typing-inspection==0.4.1
+    # via pydantic
 urllib3==2.5.0
     # via
     #   requests
@@ -157,7 +151,7 @@ virtualenv==20.34.0
     #   asv
     #   pre-commit
     #   tox
-whitenoise==6.9.0
+whitenoise==6.11.0
     # via -r requirements-dev.in
 zipp==3.23.0
     # via importlib-metadata
