+++ src/django_components/util/exception.py
@@ -20,19 +20,26 @@ def component_error_message(component_path: List[str]) -> Generator[None, None, 
         components = getattr(err, "_components", [])
         components = err._components = [*component_path, *components]  # type: ignore[attr-defined]
 
-        # Access the exception's message, see https://stackoverflow.com/a/75549200/9788634
-        if len(err.args) and err.args[0] is not None:
-            if not components:
-                orig_msg = str(err.args[0])
-            else:
-                orig_msg = str(err.args[0]).split("\n", 1)[-1]
-        else:
-            orig_msg = str(err)
-
         # Format component path as
         # "MyPage > MyComponent > MyComponent(slot:content) > Base(slot:tab)"
         comp_path = " > ".join(components)
         prefix = f"An error occured while rendering components {comp_path}:\n"
+
+        # Access the exception's message, see https://stackoverflow.com/a/75549200/9788634
+        if len(err.args) and err.args[0] is not None:
+            orig_msg = str(err.args[0])
+            if components and "An error occured while rendering components" in orig_msg:
+                orig_msg = str(err.args[0]).split("\n", 1)[-1]
+        else:
+            # When the exception has no message, it may be that the exception
+            # does NOT rely on the `args` attribute. Such case is for example
+            # Pydantic exceptions.
+            #
+            # In this case, we still try to use the `args` attribute, but
+            # it's not guaranteed to work. So we also print out the component
+            # path ourselves.
+            print(prefix)  # noqa: T201
+            orig_msg = str(err)
 
         err.args = (prefix + orig_msg,)  # tuple of one
