+++ tox.ini
@@ -39,6 +39,7 @@ deps =
   # Othrwise we get error:
   # playwright._impl._errors.Error: BrowserType.launch: Executable doesn't exist at /home/runner/.cache/ms-playwright/chromium-1140/chrome-linux/chrome
   playwright==1.48.0
+  pydantic
   requests
   types-requests
   whitenoise
@@ -60,6 +61,7 @@ deps =
   syrupy  # snapshot testing
   # NOTE: Keep playwright in sync with the version in requirements-ci.txt
   playwright==1.48.0
+  pydantic
   requests
   types-requests
   whitenoise
